moves styles to styles folder and updates imports
