+++ hackathon-fe/src/pages/UpdateFriendPage/UpdateFriendPage.js
@@ -10,7 +10,7 @@ import "react-toastify/dist/ReactToastify.css";
 
 import Header from "../../components/Header/Header";
 
-import styles from "../CreateFriendPage/CreateFriendPage.module.css";
+import styles from "../../styles/CreateFriendPage.module.css";
 import { Box, MenuItem, Select } from "@mui/material";
 
 function UpdateFriendPage () {
