+++ hackathon-fe/src/styles/Footer.module.css
